move build_preview_test from commands/permeable to integration shard
