+++ packages/flutter_tools/lib/src/commands/build_preview.dart
@@ -19,13 +19,13 @@ import 'build.dart';
 class BuildPreviewCommand extends BuildSubCommand {
   BuildPreviewCommand({
     required super.logger,
-    required bool verboseHelp,
+    required super.verboseHelp,
     required this.fs,
     required this.flutterRoot,
     required this.processUtils,
     required this.artifacts,
-  }) : super(verboseHelp: verboseHelp) {
-    addCommonDesktopBuildOptions(verboseHelp: verboseHelp);
+  }) {
+    //addCommonDesktopBuildOptions(verboseHelp: verboseHelp);
   }
 
   @override
